Perma-locked slots fix.
This system had a fatal flaw that would quite often lead to slots becoming locked indefinitely. I was especially seeing this on bard, with rings like warp rings, sometimes with ranged ammo pouches, and others.  This revision fixes the pitfalls.
Essentially what was happening was set_combine() was getting called with nil and trying to merge a missing entry  and that would lead to slots staying disabled and gearswap never looking at them again until reloaded or unloaded entirely and reloaded.
 i've tested it quite a bit already and so far, super clean. Everything works exactly as intended now.